+++ modules/Material/ApplicationWindow.qml
@@ -101,7 +101,8 @@ Controls.ApplicationWindow {
         onPopped: __toolbar.pop(  )
     }
 
-    Item {
+
+    Rectangle {
         id: overlayLayer
         objectName: "overlayLayer"
 
@@ -109,12 +110,32 @@ Controls.ApplicationWindow {
         z: 100
 
         property Item currentOverlay
+        color: "transparent"
+
+        states: State {
+            name: "ShowState"
+            when: overlayLayer.currentOverlay != null
+
+            PropertyChanges {
+                target: overlayLayer
+                color: currentOverlay.backdropColor
+            }
+        }
+
+        transitions: Transition {
+            ColorAnimation {
+                duration: 300
+                easing.type: Easing.InOutQuad
+            }
+
+        }
 
         MouseArea {
             anchors.fill: parent
             enabled: overlayLayer.currentOverlay != null
             hoverEnabled: enabled
             onClicked: overlayLayer.currentOverlay.close()
+            onWheel: wheel.accepted = true
         }
     }