+++ CHANGELOG.md
@@ -3,4 +3,4 @@ QML Material CHANGELOG
 
 
  * Hide the CSD window controls when client-side decorations are disabled
- 
+ * Improve the dialog component to better follow Material Design